requirements.txt
requests==2.32.5
dotenv==0.9.9
fitz==0.0.1.dev2
PyMuPDF==1.26.4
bs4==0.0.2
faiss-cpu==1.12.0
numpy==2.2.6
scipy==1.15.3
sentence-transformers==5.1.0
FlagEmbedding==1.3.5
transformers==4.56.2
langchain==0.3.27
langchain_community==0.3.29
pydantic== 2.11.9
langchain_openai==0.3.33
langgraph==0.6.7
dependency_injector==4.48.2
langsmith==0.4.29
IPython==8.37.0